+++ CHANGELOG.md
@@ -1,5 +1,29 @@
-## [Unreleased]
 
-## [0.1.0] - 2025-07-19
+# keycloak Middleware Gem
 
-- Initial release
+This gem add Keycloak authentication middleware to your Rails app, protect specific paths, and enforce roles.
+
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+This project adheres to [Semantic Versioning](https://semver.org/).
+
+---
+
+## [Released]
+
+### Added
+- Initial release of `KeycloakMiddleware`
+- Plug-and-play Rails middleware for Keycloak authentication
+- Configurable protected paths and role-based access control
+- JWKS fetching and JWT validation
+- Rails generator to create initializer
+- Example `.env` file
+
+---
+
+## [0.1.4] - 2025-07-19
+
+### Added
+- First stable release.